@@ -1,7 +1,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0">
     <modelVersion>4.0.0</modelVersion>
     <groupId>com.fonarik94</groupId>
-    <artifactId>PhotoViewer 1.0</artifactId>
+    <artifactId>PhotoViewer</artifactId>
     <packaging>jar</packaging>
     <version>0.1-SNAPSHOT</version>
     <build>
@@ -10,8 +10,8 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <configuration>
-                    <source>1.7</source>
-                    <target>1.7</target>
+                    <source>1.8</source>
+                    <target>1.8</target>
                 </configuration>
             </plugin>
         </plugins>
